% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{geocode_address}
\alias{geocode_address}
\title{Retrieve the call center data dataset}
\usage{
geocode_address(batch, fields)
}
\arguments{
\item{batch}{the data.frame}

\item{fields}{the character vector of the address}
}
\value{
A dataframe.
}
\description{
\code{geocode_address} this function runs address strings through the city of Milwaukee's
address-then-DIME geocoder
}
