% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_DelinquentRealEstate.R
\name{get_DelinquentRealEstate}
\alias{get_DelinquentRealEstate}
\title{Retrieve a list of delinquent real estate
\code{get_DelinquentRealEstate} returns a data.frame containing all delinquent real estate
accounts geocoded (if specified) and filtered for the selected geography
  (if specified).}
\usage{
get_DelinquentRealEstate(shape, spatial = FALSE,
  include_missing = FALSE)
}
\arguments{
\item{shape}{An object of class sf. If included, the output will be filtered using
st_intersection}

\item{spatial}{Logical. If TRUE the output is class sf. Defaults to FALSE.}

\item{include_missing}{Logical. If TRUE values not geocoded will be added to the output.
Defaults to FALSE.}
}
\value{
A dataframe.
}
\description{
Refer to the data dictionary for variable descriptions:
  \url{https://data.milwaukee.gov/dataset/delinquent-real-estate-tax-accounts}
}
\examples{
get_DelinquentRealEstate()
get_DelinquentRealEstate(spatial = TRUE)
}
