% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_safety_functions.R
\name{get_ems}
\alias{get_ems}
\title{Retrieve the Emergency Calls for Service (EMS) dataset}
\usage{
get_ems(start_date, end_date, include_missingDate)
}
\arguments{
\item{start_date}{The first date to be included. Must be coercible to class Date.
Defaults to first date available.}

\item{end_date}{The last date to be included. Must be coercible to class Date.
Defaults to last date available.}

\item{include_missingDate}{Logical. If TRUE values without a valid date will be included. Defaults to FALSE.}
}
\value{
A dataframe.
}
\description{
\code{get_ems} returns a data.frame containing a list of fire dept incidents
  for the requested time period (if specified).
}
\details{
No geocoding is available as the only geographies provided are zip codes and aldermanic districts. Refer to the data dictionary for variable descriptions:
  \url{https://data.milwaukee.gov/dataset/mfdems}
}
\examples{
get_ems()
get_ems(start_date = as.Date("2018-01-01"), end_date = as.Date("2018-02-01"))
}
