% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_safety_functions.R
\name{get_FireIncidents}
\alias{get_FireIncidents}
\title{Retrieve the Fire Incident dataset}
\usage{
get_FireIncidents(start_date, end_date, spatial, shape)
}
\arguments{
\item{start_date}{The first date to be included. Must be coercible to class Date.
Defaults to first date available.}

\item{end_date}{The last date to be included. Must be coercible to class Date.
Defaults to last date available.}

\item{shape}{An object of class sf. If included, the output will be filtered using
st_intersection}

\item{make_spatial}{Logical. If TRUE the output is class sf. Defaults to FALSE.}
}
\value{
A dataframe.
}
\description{
\code{get_FireIncidents} returns a data.frame containing a list of fire dept incidents
  for the requested time period (if specified).
}
\details{
Geocoding is based on the coordinates provided in the file. Refer to the data dictionary for variable descriptions:
  \url{https://data.milwaukee.gov/dataset/mfdprimary/}
}
\examples{
get_FireIncidents()
get_FireIncidents(start_date = as.Date("2018-01-01"), end_date = as.Date("2018-02-01"))
}
