% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food_and_drink_functions.R, R/get_FoodGrades.R
\name{get_FoodGrades}
\alias{get_FoodGrades}
\title{Retrieve a list of current food establishment grades from the health department}
\usage{
get_FoodGrades(spatial = FALSE, shape, include_missing = FALSE)

get_FoodGrades(spatial = FALSE, shape, include_missing = FALSE)
}
\arguments{
\item{spatial}{Logical. If TRUE the output is class sf. Defaults to FALSE.}

\item{shape}{An object of class sf. If included, the output will be filtered using
st_intersection}

\item{include_missing}{Logical. If TRUE values not geocoded will be added to the output.
Defaults to FALSE.}

\item{spatial}{Logical. If TRUE the output is class sf. Defaults to FALSE.}

\item{shape}{An object of class sf. If included, the output will be filtered using
st_intersection}

\item{include_missing}{Logical. If TRUE values not geocoded will be added to the output.
Defaults to FALSE.}
}
\value{
A dataframe.

A dataframe.
}
\description{
\code{get_FoodGrades} returns a data.frame containing all current restaurant grades
geocoded (if specified) and filtered for the selected geography
  (if specified).

\code{get_FoodGrades} returns a data.frame containing all current restaurant grades
geocoded (if specified) and filtered for the selected geography
  (if specified).
}
\details{
Refer to the data dictionary for variable descriptions:
  \url{https://data.milwaukee.gov/dataset/foodgrading}

Refer to the data dictionary for variable descriptions:
  \url{https://data.milwaukee.gov/dataset/foodgrading}
}
\examples{
get_FoodGrades()
get_FoodGrades(spatial = TRUE)
get_FoodGrades()
get_FoodGrades(spatial = TRUE)
}
