% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_services_functions.R
\name{get_GarbageViolations}
\alias{get_GarbageViolations}
\title{Retrieve the resolved (abated) garbage violations dataset}
\usage{
get_GarbageViolations(start_date, end_date, shape, include_missing,
  spatial)
}
\arguments{
\item{start_date}{The first date to be included. Must be coercible to class Date.
Defaults to first date available.}

\item{end_date}{The last date to be included. Must be coercible to class Date.
Defaults to last date available.}

\item{shape}{An object of class sf. If included, the output will be filtered using
st_intersection}

\item{include_missing}{Logical. If TRUE values not geocoded will be added to the output. Defaults to FALSE.}

\item{spatial}{Logical. If TRUE the output is class sf. Defaults to FALSE.}
}
\value{
A dataframe.
}
\description{
\code{get_GarbageViolations} returns a data.frame containing the complete Garbage Violations dataset
  for the requested time period (if specified) and the selected geography
  (if specified).
}
\details{
Refer to the data dictionary for variable descriptions:
  \url{https://data.milwaukee.gov/dataset/garbageviolationsabatedcurrent}
}
\examples{
get_GarbageViolations()
get_GarbageViolations(start_date = as.Date("2018-01-01"), end_date = as.Date("2018-02-01"))
}
