% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography_functions.R
\name{get_geography}
\alias{get_geography}
\title{Retrieve a specified geography as a simple features object}
\usage{
get_geography(name)
}
\arguments{
\item{name}{The name of the desired geography. Options are:
\describe{
\item{police.districts}{\href{https://data.milwaukee.gov/dataset/milwaukee-police-district}{Milwaukee Police Department districts}}
\item{mpd.squad.areas}{\href{https://data.milwaukee.gov/dataset/milwaukee-police-department-squad-areas}{Milwaukee Police Department squad areas}}
\item{neighborhoods}{the shapes of neighborhoods defined by the \href{https://data.milwaukee.gov/dataset/neighborhoods}{Department of City Development}}
\item{mpd.reporting.districts}{\href{https://data.milwaukee.gov/dataset/milwaukee-police-department-reporting-districts}{Milwaukee Police Department reporting units}}
\item{voting.wards}{\href{https://data.milwaukee.gov/dataset/voting-wards}{Voting wards}}
\item{redevelopment.plans}{\href{https://data.milwaukee.gov/dataset/redevelopment-plans}{redevelopment plans}}
\item{snow.plowing.areas}{\href{https://data.milwaukee.gov/dataset/snow-removal}{Snow removal routes and plowing areas}}
\item{streetcar.stops}{\href{https://data.milwaukee.gov/dataset/street-car-stops}{Locations of streetcar stops}}
\item{city.maintained.land}{\href{https://data.milwaukee.gov/dataset/city-maintained-parcels}{Parcel outlines of city-owned properties}}
\item{garbage.routes}{\href{https://data.milwaukee.gov/dataset/garbage-recycling-collection}{Garbage routes}}
\item{recycling.summer.routes}{\href{https://data.milwaukee.gov/dataset/garbage-recycling-collection}{Summer recycling routes}}
\item{recycling.winter.routes}{\href{https://data.milwaukee.gov/dataset/garbage-recycling-collection}{Winter recycling routes}}
\item{sanitation.districts}{\href{https://data.milwaukee.gov/dataset/garbage-recycling-collection}{Sanitation districts}}
\item{street.sweeping}{\href{https://data.milwaukee.gov/dataset/street-sweeping}{Street sweeping routes}}
\item{bids}{\href{https://data.milwaukee.gov/dataset/business-improvement-districts-bid}{Business Improvement Districts}}
\item{strategic.action.plans}{\href{https://data.milwaukee.gov/dataset/strategic-action-plans}{Strategic action plans}}
\item{adult.family.homes}{\href{https://data.milwaukee.gov/dataset/community-based-residential-facilities-within-city-of-milwaukee}{Adult family homes}}
\item{leaf.collection}{\href{https://data.milwaukee.gov/dataset/leaf-collection}{Leaf collection areas}}
\item{aldermanic.districts}{\href{https://data.milwaukee.gov/dataset/aldermanic-districts}{Outlines of aldermanic districts}}
\item{child.group.homes}{\href{https://data.milwaukee.gov/dataset/child-residential-facilities-within-city-of-milwaukee}{Locations of child residential facilities}}
\item{liquor.licenses}{\href{https://data.milwaukee.gov/dataset/alcohol-licenses}{Parcel centroids associated with liquor licenses}}
\item{schools}{\href{https://data.milwaukee.gov/dataset/schools-elementary-secondary}{Parcels identified as elementary or secondary schools}}

}

Defaults to first date available.}
}
\value{
A dataframe of class sf.
}
\description{
\code{get_geography} returns a data.frame containing the complete WIBRS crime data
  for the requested time period (if specified) and the selected geography
  (if specified).
}
\details{
Refer to the linked descriptions for more information:
}
\examples{
get_geography(name = "neighborhoods")
}
