% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housing_and_property_functions.R
\name{get_ImprovedSales}
\alias{get_ImprovedSales}
\title{\code{get_ImprovedSales} returns a data.frame containing addresses of every foreclosed
property sold by the city. The data.frame can be geocoded
 (if specified) and filtered for the selected geography (if specified).}
\usage{
get_ImprovedSales(start_year, end_year, shape, spatial, include_missing)
}
\arguments{
\item{start_year}{a numeric value. Defaults to earliest available}

\item{end_year}{a numeric value. Defaults to last available.}

\item{shape}{An object of class sf. If included, the output will be filtered using
st_intersection}

\item{spatial}{Logical. If TRUE the output is class sf. Defaults to FALSE.}

\item{include_missing}{Logical. If TRUE values not geocoded will be added to the output.
Defaults to FALSE.}
}
\value{
A dataframe.
}
\description{
Refer to the data dictionary for further information:
  \url{https://data.milwaukee.gov/dataset/total-number-of-city-owned-improved-properties-sold-by-year}
}
\examples{
get_ImprovedSales()
get_ImprovedSales(spatial = TRUE)
}
