% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housing_and_property_functions.R
\name{get_LeadService}
\alias{get_LeadService}
\title{\code{get_LeadService} returns a data.frame containing addresses in which the city-owned
section of the water service line is made of lead. The data.frame can be geocoded
 (if specified) and filtered for the selected geography (if specified).}
\usage{
get_LeadService(shape, spatial, include_missing)
}
\arguments{
\item{shape}{An object of class sf. If included, the output will be filtered using
st_intersection}

\item{spatial}{Logical. If TRUE the output is class sf. Defaults to FALSE.}

\item{include_missing}{Logical. If TRUE values not geocoded will be added to the output.
Defaults to FALSE.}
}
\value{
A dataframe.
}
\description{
Refer to the data dictionary for further information:
  \url{https://data.milwaukee.gov/dataset/lead-service-line-data}
}
\examples{
get_LeadService()
get_LeadService(spatial = TRUE)
}
