% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food_and_drink_functions.R
\name{get_LiquorLicenses}
\alias{get_LiquorLicenses}
\title{Retrieve a list of current liquor licenses}
\usage{
get_LiquorLicenses(spatial, shape, include_missing)
}
\arguments{
\item{spatial}{Logical. If TRUE the output is class sf. Defaults to FALSE.}

\item{shape}{An object of class sf. If included, the output will be filtered using
st_intersection}

\item{include_missing}{Logical. If TRUE values not geocoded will be added to the output.
Defaults to FALSE.}
}
\value{
A dataframe.
}
\description{
\code{get_LiquorLicenses} returns a data.frame containing all current liquor licenses
geocoded (if specified) and filtered for the selected geography
  (if specified).
}
\details{
Refer to the data dictionary for variable descriptions:
  \url{https://data.milwaukee.gov/dataset/liquorlicenses}
}
\examples{
get_LiquorLicenses()
get_LiquorLicenses(spatial = TRUE)
}
