% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housing_and_property_functions.R
\name{get_mai}
\alias{get_mai}
\title{Retrieve the Master Address Index (MAI)}
\usage{
get_mai(spatial, shape, include_missing)
}
\arguments{
\item{spatial}{Logical. If TRUE the output is class sf. Defaults to FALSE.}

\item{shape}{An object of class sf. If included, the output will be filtered using
st_intersection}

\item{include_missing}{Logical. If TRUE values not geocoded will be added to the output.
Defaults to FALSE.}
}
\value{
A dataframe.
}
\description{
\code{get_mai} returns a data.frame containing the complete Master Address index
geocoded (if specified) and filtered for the selected geography
  (if specified).
}
\details{
Refer to the data dictionary for variable descriptions:
  \url{https://data.milwaukee.gov/dataset/mai}
}
\examples{
get_mai()
get_mai(spatial = TRUE)
}
