% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mprop_function.R
\name{get_mprop}
\alias{get_mprop}
\title{Retrieve the Master Property List (MPROP)}
\usage{
get_mprop(start_year, end_year)
}
\arguments{
\item{start_year}{The first year to be included.}

\item{end_year}{The last year to be included.}
}
\value{
A dataframe.
}
\description{
\code{get_mprop} returns a data.frame containing the complete MPROP file
  for each specified year. The column "year" indicates the MPROP vintage.
  Currently only the years 2016 to 2018 are available. 2018 data is updated
  daily.
}
\details{
Refer to the data dictionary for variable descriptions:
  \url{https://data.milwaukee.gov/dataset/mprop}
}
\examples{
get_mprop(start_year = 2016, end_year = 2018)
get_mprop(start_year = 2018, end_year = 2018)
}
