% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elections_functions.R, R/get_PollingPlace.R
\name{get_PollingPlace}
\alias{get_PollingPlace}
\title{Retrieve the call center data dataset}
\usage{
get_PollingPlace(shape, spatial = FALSE)

get_PollingPlace(shape, spatial = FALSE)
}
\arguments{
\item{shape}{An object of class sf. If included, the output will be filtered using
st_intersection}

\item{spatial}{Logical. If TRUE the output is class sf. Defaults to FALSE.}

\item{include_missing}{Logical. If TRUE values not geocoded will be added to the output. Defaults to FALSE.}

\item{spatial}{Logical. If TRUE the output is class sf. Defaults to FALSE.}

\item{shape}{An object of class sf. If included, the output will be filtered using
st_intersection}

\item{include_missing}{Logical. If TRUE values not geocoded will be added to the output. Defaults to FALSE.}
}
\value{
A dataframe.

A dataframe.
}
\description{
\code{get_PollingPlace} returns a data.frame containing information about polling places
  in the city of Milwaukee for the selected geography (if specified).

\code{get_PollingPlace} returns a data.frame containing information about polling places
  in the city of Milwaukee for the selected geography (if specified).
}
\details{
Refer to the data dictionary for variable descriptions:
  \url{https://data.milwaukee.gov/dataset/votingward}

Refer to the data dictionary for variable descriptions:
  \url{https://data.milwaukee.gov/dataset/votingward}
}
\examples{
get_PollingPlace()
get_PollingPlace(spatial = TRUE)
get_PollingPlace()
get_PollingPlace(spatial = TRUE)
}
