% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_TrafficAccidents.R
\name{get_TrafficAccidents}
\alias{get_TrafficAccidents}
\title{Retrieve the Traffic Accident dataset}
\usage{
get_TrafficAccidents(start_date = NULL, end_date = NULL,
  include_missingDate = FALSE)
}
\arguments{
\item{start_date}{The first date to be included. Must be coercible to class Date.
Defaults to first date available.}

\item{end_date}{The last date to be included. Must be coercible to class Date.
Defaults to last date available.}

\item{include_missingDate}{Logical. If TRUE values without a valid date will be included. Defaults to FALSE.}
}
\value{
A dataframe.
}
\description{
\code{get_TrafficAccidents} returns a data.frame containing a list of traffic accidents
  for the requested time period (if specified).
}
\details{
Currently, geocoding is not available
  for this resource because locations are not given as standard addresses. Refer to the data dictionary for variable descriptions:
  \url{https://data.milwaukee.gov/dataset/trafficaccident}
}
\examples{
get_TrafficAccidents()
get_TrafficAccidents(start_date = as.Date("2018-01-01"), end_date = as.Date("2018-02-01"))
}
