% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_WorkPermits.R
\name{get_WorkPermits}
\alias{get_WorkPermits}
\title{Retrieve a list of current residential and commercial work permits}
\usage{
get_WorkPermits(start_date = NULL, end_date = NULL, spatial = FALSE,
  shape, include_missing = FALSE)
}
\arguments{
\item{start_date}{must be coercible to date format}

\item{end_date}{must be coercible to date format}

\item{spatial}{Logical. If TRUE the output is class sf. Defaults to FALSE.}

\item{shape}{An object of class sf. If included, the output will be filtered using
st_intersection}

\item{include_missing}{Logical. If TRUE values not geocoded will be added to the output.
Defaults to FALSE.}
}
\value{
A dataframe.
}
\description{
\code{get_WorkPermits} returns a data.frame containing all residential and commercial
work permits geocoded (if specified) and filtered for the selected geography
  (if specified).
}
\details{
Refer to the data dictionary for variable descriptions:
  \url{https://data.milwaukee.gov/dataset/liquorlicenses}
}
\examples{
get_WorkPermits()
get_WorkPermits(spatial = TRUE)
}
